ai_service/requirements.txt
numpy
pandas
chromadb
open-clip-torch
langchain
langchain-openai
mysql-connector-python
python-dotenv
openai
fastapi
uvicorn
pydantic
datasets
sentence-transformers
